desktop grid: use -/+ text buttons
esp. the list-remove icon looks like "delete" and
can cause worries about what happens to the windows

as a bonus, the buttons now follow the DPI. are a
little smaller (presently might fit touch devices? but
is way to huge on "normal" desktops with ordinary DPI)
and the buttons have nice visible animations on pressing

BUG: 354131
FIXED-IN: 5.5