Only focusToNull when activating a ShellClient
ShellClient::doSetActive calls into focusToNull to deactivate active
windows on X11. But this should only be done when activating a
ShellClient. When deactivating a ShellClient the focusToNull can
result in an activated X-client to lose activation.

Thus this change ensures that focusToNull is only performed for a
now active ShellClient.

Reviewed-By: bshah