Call GLRenderTarget::setVirtualScreenGeometry prior to self test
Summary:
The shader self test uses GLRenderTarget and restores the viewport. Due
to the setVirtualScreenGeometry not being set the calculated viewport is
incorrect and any rendering breaks. This is a regression introduced with
1708278a451e6dab92549f9cabe6ad13ea3b4ba2 which only affects Wayland
backends which do not perform per screen rendering.

The glxbackend was not affected as it sets the viewport on screen
geometry change.

Reviewers: bshah, #plasma_on_wayland, #kwin

Reviewed By: bshah, #plasma_on_wayland, #kwin

Subscribers: plasma-devel, kwin

Tags: #plasma_on_wayland, #kwin

Differential Revision: https://phabricator.kde.org/D3155