Pass a desktopfile name to kcmshell5
Summary:
In order to have a proper window icon for kcmshell5 on Wayland KWin
needs to pass a desktop file name as command line argument.

The desktop file of kwinfocus is used as it uses the
preferences-system-windows icon which is also the icon on X11.

Reviewers: #kwin, #plasma_on_wayland

Subscribers: plasma-devel, kwin

Tags: #plasma_on_wayland, #kwin

Differential Revision: https://phabricator.kde.org/D3162