Fix the osd re-placement on X11
Summary:
6383f8304711719fe22371474138efc144adc750 introduced a regression hitting
an assert on X11 when an osd window is added.

This change tries to address this by adding a GeometryUpdateBlocker
around the change. The hit assertion is:

    pendingGeometryUpdate() == PendingGeometryNone ||
    areGeometryUpdatesBlocked()

By using a GeometryUpdateBlocker the second condition becomes true
and we no longer hit the condition.

BUG: 374038

Test Plan:
Manual tested both on X11 and Wayland, Wayland autotest
still passes

Reviewers: #kwin, #plasma

Subscribers: plasma-devel, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D3808