Correct inital loading of BorderActivate
Summary:
We have a comma separated list:

"".split(',') returns ""
JS decides that's worth itterating over, and we implicitly cast "" to a
number which means we register on screen edge 0 (the top).

We can't use typeof or isFinite because valid entries are still strings
at this point.

Test Plan:
Ran with default config, no longer registered on top edge
Set an edge properly in the KCM. Still worked.

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: mart, plasma-devel, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D4113