[autotests] Ensure TestDontCrashUseractionsMenu uses breeze
The test introduced with 962a2e39ee7a476e416f067dd31f10e3a4a63784 is
meant to be used with the breeze widget style as that one triggered the
crashy code. Given the test coverage build.kde.org does not run with the
breeze widget style, so enforce this through manual loading.