Remove hard X11 runtime dependency from DBusInterface
Summary:
KWin's DBusInterface announces the service name on an X11 root window
property. For this it requires a runtime dependency on X11. If we would
try to start kwin_wayland without XWayland support this would result in
a crash.

This change turns the dependency into an optional one by checking
whether the KWin::Application has support for X11 and registers a
connection to announce the property as soon as a connection is
established. Thus it even supports the restart of XWayland now.

Test Plan: Compiles

Reviewers: #kwin, #plasma

Subscribers: plasma-devel, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D7474