Disable synced resizing for X11 clients on Xwayland
Summary:
The sync is broken due to updateXTime not working correctly on Wayland.
As the broken sync results in a very unsmooth resize experience it is
better to disable it for the time being.

Once we have a solution for updateXTime it should be enabled again.

BUG: 374881

Test Plan: Run nested kwin_wayland and resized Qt 4 application

Reviewers: #kwin, #plasma

Subscribers: plasma-devel, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D7524